+++ GenericModConfigMenu/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* Improved mod manifest validation.
+
 ## 1.4.0
 Released 10 July 2021 for Stardew Valley 1.5. Updated by Pathoschild.